+++ src/about-menu/open-about-menu.ts
@@ -3,7 +3,7 @@ import path from 'path';
 import * as Sentry from '@sentry/electron/renderer';
 import { sprintf, __ } from '@wordpress/i18n';
 import { ABOUT_WINDOW_HEIGHT, ABOUT_WINDOW_WIDTH } from '../constants';
-import { shellOpenExternalWrapper } from '../utils/shellOpenExternalWrapper';
+import { shellOpenExternalWrapper } from '../utils/shell-open-external-wrapper';
 
 let aboutWindow: BrowserWindow | null = null;
